queryDB/requirements.txt
psycopg2-binary==2.9.9
google-generativeai==0.8.3
asyncio